+++ components/grid/grid-helpers/index.js
@@ -1,0 +1,3 @@
+export { GridColumn } from './grid-column';
+export * from './grid-helpers';
+export { TreeGroupColumn } from './tree-grid-group-column';
