+++ components/grid/index.js
@@ -1,5 +1,5 @@
 import './custom.scss';
 export { PaginatedGrid } from './paginated-grid';
-export { GridColumn } from './grid-column';
+export { GridColumn } from './grid-helpers';
 export { SimpleGrid } from './simple-grid';
 export { TreeGrid } from './tree-grid';
